requirements.txt
Flask==2.1.2
Flask-Dance[sqla]==6.0.0
blinker==1.4
python-dotenv==0.18.0

PyMySQL~=1.0.2